restructure mailchimp library to allow for a list class
